Refactor in ART  API  - sub app layer
