+++ ART/plugin/sub_app_module/fermat-art-plugin-sub-app-module-fan-community-bitdubai/src/main/java/com/bitdubai/fermat_art_plugin/layer/sub_app_module/fan_community/developer/bitdubai/version_1/structure/FanCommunityManager.java
@@ -17,7 +17,7 @@ import com.bitdubai.fermat_api.layer.modules.exceptions.CantGetSelectedActorIden
 import com.bitdubai.fermat_art_api.layer.actor_connection.fan.interfaces.FanActorConnectionManager;
 import com.bitdubai.fermat_art_api.layer.actor_connection.fan.interfaces.FanActorConnectionSearch;
 import com.bitdubai.fermat_art_api.layer.actor_connection.fan.utils.FanLinkedActorIdentity;
-import com.bitdubai.fermat_art_api.layer.sub_app_module.community.FanCommunityModuleManager;
+import com.bitdubai.fermat_art_api.layer.sub_app_module.community.fan.FanCommunityModuleManager;
 import com.bitdubai.fermat_pip_api.layer.platform_service.error_manager.interfaces.ErrorManager;
 
 import java.util.UUID;
